// src/app/pages/si_sesions/components/si_sesions-table/si_sesions.interface.ts
export interface Si_sesionsInterface {
   idsi_sesion?: number;
   codigo?: string;
   descripcion?: string;
   tipo?: string;
   si_user_idsi_user?: number;
   estado?: string;
   latitude?: string;
   longitude?: string;
   accuracy?: string;
   is_deleted?: boolean;
   created_by?: number;
   created_at?: string;
   modified_at?: string;
}